Keep browser sign-in helpers private
